chore(app): remove dead code and consolidate hook directories
- Delete unused useBidSubmission hook (replaced by useEscrowBidSubmission)
- Rename ForecastPageImp to ForecastPageContent to match naming convention
- Drop unused className prop from LegacyBadge
- Consolidate hooks/relay, hooks/relayer, hooks/data into hooks/ws
- Move components/relay/PeerIndicator into components/shared
- Inline markets/forms barrel (single-export re-export)
- Strip stale "removed X" leftover comments

No functionality or UI changes.
