--- packages/app/src/components/markets/forms/index.ts
@@ -1,1 +1,0 @@
-export { PositionSizeInput } from './inputs/PositionSizeInput';
