+++ packages/app/src/hooks/auction/index.ts
@@ -1,5 +1,2 @@
 export * from './useBidPreflight';
 export * from './useEscrowBidSubmission';
-
-// Escrow auction hooks
-export * from './useBidSubmission';
